refactor: Update kotlin and dependencies versions, room database, and remove compose compiler
- Updated Kotlin version to 1.9.10
- Updated dependencies version
- Removed compose compiler plugin
- Updated Room database
- Removed unused imports
- Change id type in database to Int

+++ app/src/main/java/com/example/arcanemusic/ui/home/SongListViewModel.kt
@@ -47,7 +47,7 @@ class SongListViewModel(
                     val duration = it.getColumnIndexOrThrow(MediaStore.Audio.Media.DURATION)
 
                     while (it.moveToNext()) {
-                        val id = it.getLong(id)
+                        val id = it.getInt(id)
                         val title = it.getString(title)
                         val artist = it.getString(artist)
                         val data = it.getString(data)
